requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.3; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
alpha-vantage==2.3.1
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.2.0; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6
dnspython==2.5.0; python_version >= '3.8'
email-validator==2.1.0.post1
fastapi[all]==0.109.0; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.2; python_version >= '3.8'
httptools==0.6.1
httpx==0.26.0
idna==3.6; python_version >= '3.5'
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.4; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
numpy==1.26.3; python_version >= '3.12'
orjson==3.9.12
pandas==2.2.0; python_version >= '3.9'
pydantic==2.6.0; python_version >= '3.8'
pydantic-core==2.16.1; python_version >= '3.8'
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.6
pytz==2023.4
pyyaml==6.0.1
redis==5.0.1; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.35.1; python_version >= '3.8'
strawberry-graphql[fastapi]==0.219.1; python_version >= '3.8' and python_version < '4.0'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2023.4; python_version >= '2'
ujson==5.9.0
urllib3==2.2.0; python_version >= '3.8'
uvicorn[standard]==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4; python_version >= '3.7'